Configures environment variables for client app
Introduces environment variable configuration for the client application.

Adds an `.env.example` file to serve as a template for configuration.
Moves environment variable loading and checking to a dedicated module for better organization and reusability.
Modifies build and server tasks to utilize the new environment variable handling.
Updates the README to include instructions on configuring the environment.

Fix #43